fix bug (#198)
add test

add new test_case

fix

add test case

fix

fix

add test

fix test

fix test

fix

fix

fix

fix

+++ gluonnlp/data/batchify.py
@@ -77,9 +77,9 @@ def _stack_arrs(arrs, use_shared_mem=False):
     else:
         out = np.asarray(arrs)
         if use_shared_mem:
-            return mx.nd.array(out, ctx=mx.Context('cpu_shared', 0))
+            return mx.nd.array(out, ctx=mx.Context('cpu_shared', 0), dtype=out.dtype)
         else:
-            return mx.nd.array(out)
+            return mx.nd.array(out, dtype=out.dtype)
 
 
 class Stack(object):
